:truck: Move dataclasses/tuples to types
