+++ ice_station_zebra/types.py
@@ -1,9 +1,10 @@
 from collections.abc import Iterator, Mapping, Sequence
 from dataclasses import dataclass
 from enum import IntEnum
-from typing import Any, Self, TypedDict
+from typing import Any, Literal, NamedTuple, Self, TypedDict
 
 from jaxtyping import Float
+from matplotlib.colors import Normalize
 from numpy import float32
 from numpy.typing import NDArray
 from omegaconf import DictConfig
@@ -13,6 +14,18 @@ ArrayCHW = Float[NDArray[float32], "channels height width"]
 ArrayTCHW = Float[NDArray[float32], "time channels height width"]
 TensorNCHW = Float[Tensor, "batch channels height width"]
 TensorNTCHW = Float[Tensor, "batch time channels height width"]
+
+# DiffMode → what you compute
+# - "signed": target - prediction (can be ±, so symmetric colour scale around 0)
+# - "absolute": |target - prediction| (≥ 0, sequential scale)
+# - "smape": |pred - target| / ((|pred|+|target|)/2) (≥ 0, sequential scale)
+DiffMode = Literal["signed", "absolute", "smape"]
+
+# DiffStrategy → when you compute (for animations)
+# - "precompute": compute full diff stream once (fast playback, more RAM)
+# - "two-pass": scan once to figure the scale, compute per-frame (balanced)
+# - "per-frame": compute per-frame (low RAM, more CPU)
+DiffStrategy = Literal["precompute", "two-pass", "per-frame"]
 
 
 class TensorDimensions(IntEnum):
@@ -78,6 +91,25 @@ class DataSpace:
         )
 
 
+class DiffColourmapSpec(NamedTuple):
+    """Colour scale specification for visualising a difference panel.
+
+    Attributes:
+        norm: Normalisation object for mapping data values to colours.
+              Typically TwoSlopeNorm for signed differences (centred at 0).
+              None for non-signed modes (absolute, smape).
+        vmin: Lower bound for colour scale (used if norm is None).
+        vmax: Upper bound for colour scale (used if norm is None).
+        cmap: Name of the matplotlib colourmap to use.
+
+    """
+
+    norm: Normalize | None
+    vmin: float | None
+    vmax: float | None
+    cmap: str
+
+
 @dataclass
 class ModelTestOutput(Mapping[str, Tensor]):
     """Output of a model test step."""
@@ -106,3 +138,21 @@ class ModelTestOutput(Mapping[str, Tensor]):
     def __len__(self) -> int:
         """Return ModelTestOutput length."""
         return 3
+
+
+@dataclass(frozen=True)
+class PlotSpec:
+    variable: str
+    title_groundtruth: str = "Ground Truth"
+    title_prediction: str = "Prediction"
+    title_difference: str = "Difference"
+    n_contour_levels: int = 100
+    colourmap: str = "viridis"
+    include_difference: bool = True
+    diff_mode: DiffMode = "signed"
+    diff_strategy: DiffStrategy = "precompute"
+    selected_timestep: int = 0
+    vmin: float | None = 0.0
+    vmax: float | None = 1.0
+    colourbar_location: Literal["vertical", "horizontal"] = "vertical"
+    colourbar_strategy: Literal["shared", "separate"] = "shared"
